requirements.txt
# Core dependencies for Interactive Telegram Bot
websockets==12.0
requests==2.31.0
python-telegram-bot[job-queue]==22.5
python-dotenv==1.2.1
redis==5.0.1
psutil==7.1.2
